requirements.txt
# Core Django dependencies
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2

# API Documentation
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9

# Cache and Background Tasks  
redis==5.0.8
django-redis==5.4.0
celery==5.4.0

# Production Server
gunicorn==22.0.0
whitenoise==6.7.0

# Media handling
Pillow==10.4.0

# Development and Debugging
django-extensions==3.2.3

# Authentication and Security
djangorestframework-simplejwt==5.3.0

# Development dependencies
pytest-django==4.8.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
coverage==7.4.0
factory-boy==3.3.0
